Modify to contain only Checker tests
